check if version text exist before unlink

+++ public/view/updateSystem.php
@@ -1,5 +1,7 @@
 <?php
-unlink(PATH_HOME . "_config/updates/version.txt");
+if (file_exists(PATH_HOME . "_config/updates/version.txt"))
+    unlink(PATH_HOME . "_config/updates/version.txt");
+
 new \Config\UpdateSystem();
 ?>
 <script>
